requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
